Refatoração MaterialoControler e Testes de Contexto MaterialController

+++ src/main/java/com/abcontrol/controller/MaterialController.java
@@ -1,48 +1,80 @@
 package com.abcontrol.controller;
 
+import com.abcontrol.entity.FormaPagamentoEntity;
 import com.abcontrol.entity.MaterialEntity;
 import com.abcontrol.repository.MaterialRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 
-
+/**
+ * Validacao pela anotacao @Valid, ou seja, caso os parametros não estejam de
+ * acordo com as anotacoes de validacao nos atributos do FormaPagamentoEntity,
+ * retorna http status 400 Bad Request automaticamente, junto com a mensagem
+ * de erro de qual parametro esta incorreto.
+ */
 @RestController
 @RequestMapping(value="/material")
 public class MaterialController {
 
-    @Autowired//injeção de dependencias com
+    @Autowired
     private MaterialRepository materialRepository;
 
-    //Listar todos os matériais
     @GetMapping
-    public List<MaterialEntity> listaMateriais() {
-        return materialRepository.findAll();
+    public ResponseEntity<List<MaterialEntity>> listarMateriais() {
+
+        return ResponseEntity.ok(materialRepository.findAll());
+
     }
 
-    //Listar por ID
     @GetMapping("/{id}")
-    public MaterialEntity listaMaterialPorId(@PathVariable(value = "id") long id) {
-        return materialRepository.findById(id);
+    public ResponseEntity<MaterialEntity> listarMaterialPorId(@Valid @PathVariable(value = "id") long id) {
+
+        if(!materialRepository.existsById(id)){
+            return ResponseEntity.notFound().build();
+        } else if(materialRepository.findById(id) == null){
+            return ResponseEntity.notFound().build();
+        }
+
+        return ResponseEntity.ok(materialRepository.findById(id));
+
     }
 
-    //Cadastrar materialEntity
     @PostMapping
-    public MaterialEntity cadastraMaterial(@RequestBody MaterialEntity materialEntity) {
-        return materialRepository.save(materialEntity);
+    public ResponseEntity<MaterialEntity> cadastrarMaterial(@Valid @RequestBody MaterialEntity materialEntity) {
+
+        return ResponseEntity.ok(materialRepository.save(materialEntity));
+
     }
 
-    //Atualizar materialEntity
     @PutMapping
-    public MaterialEntity atualizaMaterial(@RequestBody MaterialEntity materialEntity){
-        return materialRepository.save(materialEntity);
+    public ResponseEntity<MaterialEntity> atualizarMaterial(@Valid @RequestBody MaterialEntity materialEntity){
+
+        if(!materialRepository.existsById(materialEntity.getId())){
+            return ResponseEntity.notFound().build();
+        }
+
+        return ResponseEntity.ok(materialRepository.save(materialEntity));
+
     }
 
-    //Excluir material por ID
     @DeleteMapping("/{id}")
-    public void excluiMaterial(@PathVariable(value = "id")long id) {
-        materialRepository.deleteById(id);
+    public ResponseEntity<MaterialEntity> excluirMaterial(@Valid @PathVariable(value = "id")long id) {
+
+        MaterialEntity materialEntityResponse;
+
+        if(!materialRepository.existsById(id)){
+            return ResponseEntity.notFound().build();
+        } else {
+            materialEntityResponse = materialRepository.findById(id);
+            materialRepository.deleteById(id);
+        }
+
+        return ResponseEntity.ok(materialEntityResponse);
+
     }
 
 }
